feat: move events to v1 subpackage
