--- src/main/java/it/pagopa/ecommerce/commons/documents/TransactionRetriedData.java
@@ -1,20 +1,0 @@
-package it.pagopa.ecommerce.commons.documents;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.Generated;
-import lombok.NoArgsConstructor;
-import org.springframework.data.mongodb.core.mapping.Document;
-
-/**
- * Data related to retry event
- */
-@Data
-@Document
-@AllArgsConstructor
-@NoArgsConstructor
-@Generated
-public class TransactionRetriedData {
-
-    private Integer retryCount;
-}
