+++ src/main/java/it/pagopa/ecommerce/commons/documents/v1/TransactionActivatedEvent.java
@@ -1,4 +1,4 @@
-package it.pagopa.ecommerce.commons.documents;
+package it.pagopa.ecommerce.commons.documents.v1;
 
 import it.pagopa.ecommerce.commons.domain.TransactionEventCode;
 import lombok.Generated;
