+++ src/main/java/it/pagopa/ecommerce/commons/documents/v1/TransactionAuthorizationCompletedEvent.java
@@ -1,4 +1,4 @@
-package it.pagopa.ecommerce.commons.documents;
+package it.pagopa.ecommerce.commons.documents.v1;
 
 import it.pagopa.ecommerce.commons.domain.TransactionEventCode;
 import lombok.NoArgsConstructor;
