+++ src/main/java/it/pagopa/ecommerce/commons/documents/v1/TransactionAuthorizationRequestData.java
@@ -1,4 +1,4 @@
-package it.pagopa.ecommerce.commons.documents;
+package it.pagopa.ecommerce.commons.documents.v1;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
