+++ src/main/java/it/pagopa/ecommerce/commons/documents/v1/TransactionClosedEvent.java
@@ -1,4 +1,4 @@
-package it.pagopa.ecommerce.commons.documents;
+package it.pagopa.ecommerce.commons.documents.v1;
 
 import it.pagopa.ecommerce.commons.domain.TransactionEventCode;
 import lombok.NoArgsConstructor;
@@ -6,21 +6,22 @@ import lombok.ToString;
 import org.springframework.data.mongodb.core.mapping.Document;
 
 /**
- * Business event generated when sending a payment receipt to a user.
+ * Business event corresponding to a transaction closure being sent. This action
+ * notifies Nodo that the transaction has been finalized.
  */
 @Document(collection = "eventstore")
 @NoArgsConstructor
 @ToString(callSuper = true)
-public final class TransactionUserReceiptAddedEvent extends TransactionEvent<Void> {
+public final class TransactionClosedEvent extends TransactionEvent<Void> {
 
     /**
      * Convenience constructor which sets the creation date to now
      *
      * @param transactionId transaction unique id
      */
-    public TransactionUserReceiptAddedEvent(
+    public TransactionClosedEvent(
             String transactionId
     ) {
-        super(transactionId, TransactionEventCode.TRANSACTION_USER_RECEIPT_ADDED_EVENT, null);
+        super(transactionId, TransactionEventCode.TRANSACTION_CLOSED_EVENT, null);
     }
 }
