+++ src/main/java/it/pagopa/ecommerce/commons/documents/v1/TransactionExpiredEvent.java
@@ -1,4 +1,4 @@
-package it.pagopa.ecommerce.commons.documents;
+package it.pagopa.ecommerce.commons.documents.v1;
 
 import it.pagopa.ecommerce.commons.domain.TransactionEventCode;
 import lombok.NoArgsConstructor;
@@ -6,22 +6,22 @@ import lombok.ToString;
 import org.springframework.data.mongodb.core.mapping.Document;
 
 /**
- * Business event corresponding to a transaction refund.
+ * Business event corresponding to a transaction expiration.
  */
 @Document(collection = "eventstore")
 @NoArgsConstructor
 @ToString(callSuper = true)
-public final class TransactionRefundedEvent extends TransactionEvent<TransactionRefundedData> {
+public final class TransactionExpiredEvent extends TransactionEvent<TransactionExpiredData> {
     /**
      * Convenience constructor which sets the creation date to now
      *
      * @param transactionId transaction unique id
      * @param data          event-specific data
      */
-    public TransactionRefundedEvent(
+    public TransactionExpiredEvent(
             String transactionId,
-            TransactionRefundedData data
+            TransactionExpiredData data
     ) {
-        super(transactionId, TransactionEventCode.TRANSACTION_REFUNDED_EVENT, data);
+        super(transactionId, TransactionEventCode.TRANSACTION_EXPIRED_EVENT, data);
     }
 }
