+++ src/main/java/it/pagopa/ecommerce/commons/documents/v1/TransactionRefundedData.java
@@ -1,4 +1,4 @@
-package it.pagopa.ecommerce.commons.documents;
+package it.pagopa.ecommerce.commons.documents.v1;
 
 import it.pagopa.ecommerce.commons.generated.server.model.TransactionStatusDto;
 import lombok.AllArgsConstructor;
@@ -8,14 +8,14 @@ import lombok.NoArgsConstructor;
 import org.springframework.data.mongodb.core.mapping.Document;
 
 /**
- * Data related to expiration event
+ * Data related to refund event
  */
 @Data
 @Document
 @AllArgsConstructor
 @NoArgsConstructor
 @Generated
-public class TransactionExpiredData {
+public class TransactionRefundedData {
 
-    private TransactionStatusDto statusBeforeExpiration;
+    private TransactionStatusDto statusBeforeRefunded;
 }
