+++ src/main/java/it/pagopa/ecommerce/commons/documents/v1/TransactionRetriedData.java
@@ -1,6 +1,5 @@
-package it.pagopa.ecommerce.commons.documents;
+package it.pagopa.ecommerce.commons.documents.v1;
 
-import it.pagopa.ecommerce.commons.generated.server.model.TransactionStatusDto;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.Generated;
@@ -8,14 +7,14 @@ import lombok.NoArgsConstructor;
 import org.springframework.data.mongodb.core.mapping.Document;
 
 /**
- * Data related to refund event
+ * Data related to retry event
  */
 @Data
 @Document
 @AllArgsConstructor
 @NoArgsConstructor
 @Generated
-public class TransactionRefundedData {
+public class TransactionRetriedData {
 
-    private TransactionStatusDto statusBeforeRefunded;
+    private Integer retryCount;
 }
