+++ src/main/java/it/pagopa/ecommerce/commons/documents/v1/TransactionUserReceiptAddedEvent.java
@@ -1,4 +1,4 @@
-package it.pagopa.ecommerce.commons.documents;
+package it.pagopa.ecommerce.commons.documents.v1;
 
 import it.pagopa.ecommerce.commons.domain.TransactionEventCode;
 import lombok.NoArgsConstructor;
@@ -6,22 +6,21 @@ import lombok.ToString;
 import org.springframework.data.mongodb.core.mapping.Document;
 
 /**
- * Business event corresponding to a transaction closure being sent. Closing a
- * transaction notifies Nodo that the transaction has been finalized.
+ * Business event generated when sending a payment receipt to a user.
  */
 @Document(collection = "eventstore")
 @NoArgsConstructor
 @ToString(callSuper = true)
-public final class TransactionClosureFailedEvent extends TransactionEvent<Void> {
+public final class TransactionUserReceiptAddedEvent extends TransactionEvent<Void> {
 
     /**
      * Convenience constructor which sets the creation date to now
      *
      * @param transactionId transaction unique id
      */
-    public TransactionClosureFailedEvent(
+    public TransactionUserReceiptAddedEvent(
             String transactionId
     ) {
-        super(transactionId, TransactionEventCode.TRANSACTION_CLOSURE_FAILED_EVENT, null);
+        super(transactionId, TransactionEventCode.TRANSACTION_USER_RECEIPT_ADDED_EVENT, null);
     }
 }
