+++ src/main/java/it/pagopa/ecommerce/commons/domain/EmptyTransaction.java
@@ -1,6 +1,6 @@
 package it.pagopa.ecommerce.commons.domain;
 
-import it.pagopa.ecommerce.commons.documents.TransactionActivatedEvent;
+import it.pagopa.ecommerce.commons.documents.v1.TransactionActivatedEvent;
 import lombok.EqualsAndHashCode;
 
 import java.time.ZonedDateTime;
@@ -13,9 +13,8 @@ import java.util.stream.Collectors;
  * to.
  * </p>
  * <p>
- * To this class you can apply a
- * {@link it.pagopa.ecommerce.commons.documents.TransactionActivatedEvent} (and
- * get a {@link it.pagopa.ecommerce.commons.domain.TransactionActivated})
+ * To this class you can apply a {@link TransactionActivatedEvent} (and get a
+ * {@link it.pagopa.ecommerce.commons.domain.TransactionActivated})
  * </p>
  *
  * @see Transaction
@@ -43,12 +42,14 @@ public final class EmptyTransaction implements Transaction {
         );
     }
 
-    /** {@inheritDoc} */
+    /**
+     * {@inheritDoc}
+     */
     @Override
-	public Transaction applyEvent(Object event) {
-		return switch (event) {
-			case TransactionActivatedEvent transactionActivatedEvent -> this.applyActivation(transactionActivatedEvent);
-			case null, default -> this;
-		};
-	}
+    public Transaction applyEvent(Object event) {
+        return switch (event) {
+            case TransactionActivatedEvent transactionActivatedEvent -> this.applyActivation(transactionActivatedEvent);
+            case null, default -> this;
+        };
+    }
 }
