+++ src/main/java/it/pagopa/ecommerce/commons/domain/TransactionExpired.java
@@ -1,7 +1,7 @@
 package it.pagopa.ecommerce.commons.domain;
 
-import it.pagopa.ecommerce.commons.documents.TransactionExpiredEvent;
-import it.pagopa.ecommerce.commons.documents.TransactionRefundedEvent;
+import it.pagopa.ecommerce.commons.documents.v1.TransactionExpiredEvent;
+import it.pagopa.ecommerce.commons.documents.v1.TransactionRefundedEvent;
 import it.pagopa.ecommerce.commons.domain.pojos.BaseTransaction;
 import it.pagopa.ecommerce.commons.domain.pojos.BaseTransactionExpired;
 import it.pagopa.ecommerce.commons.domain.pojos.BaseTransactionWithClosureError;
