+++ src/main/java/it/pagopa/ecommerce/commons/domain/TransactionUnauthorized.java
@@ -1,6 +1,6 @@
 package it.pagopa.ecommerce.commons.domain;
 
-import it.pagopa.ecommerce.commons.documents.TransactionClosureFailedEvent;
+import it.pagopa.ecommerce.commons.documents.v1.TransactionClosureFailedEvent;
 import it.pagopa.ecommerce.commons.domain.pojos.BaseTransaction;
 import it.pagopa.ecommerce.commons.domain.pojos.BaseTransactionWithCompletedAuthorization;
 import it.pagopa.ecommerce.commons.generated.server.model.TransactionStatusDto;
