+++ src/main/java/it/pagopa/ecommerce/commons/domain/pojos/BaseTransactionExpired.java
@@ -1,6 +1,7 @@
 package it.pagopa.ecommerce.commons.domain.pojos;
 
-import it.pagopa.ecommerce.commons.documents.TransactionExpiredData;
+import it.pagopa.ecommerce.commons.documents.v1.TransactionExpiredData;
+import it.pagopa.ecommerce.commons.documents.v1.TransactionExpiredEvent;
 import lombok.AccessLevel;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
@@ -13,14 +14,12 @@ import lombok.experimental.FieldDefaults;
  * </p>
  * <p>
  * All accrued data is exposed through the same structure generated by
- * {@link it.pagopa.ecommerce.commons.documents.TransactionExpiredEvent
- * TransactionExpiredEvent}, namely
- * {@link it.pagopa.ecommerce.commons.documents.TransactionExpiredData
- * TransactionExpiredData}.
+ * {@link TransactionExpiredEvent TransactionExpiredEvent}, namely
+ * {@link TransactionExpiredData TransactionExpiredData}.
  * </p>
  *
  * @see BaseTransaction
- * @see it.pagopa.ecommerce.commons.documents.TransactionExpiredEvent
+ * @see TransactionExpiredEvent
  */
 @ToString
 @EqualsAndHashCode(callSuper = true)
