+++ src/main/java/it/pagopa/ecommerce/commons/domain/pojos/BaseTransactionRefunded.java
@@ -1,6 +1,7 @@
 package it.pagopa.ecommerce.commons.domain.pojos;
 
-import it.pagopa.ecommerce.commons.documents.TransactionRefundedData;
+import it.pagopa.ecommerce.commons.documents.v1.TransactionRefundedData;
+import it.pagopa.ecommerce.commons.documents.v1.TransactionRefundedEvent;
 import lombok.AccessLevel;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
@@ -13,13 +14,12 @@ import lombok.experimental.FieldDefaults;
  * </p>
  * <p>
  * All accrued data is exposed through the same structure generated by
- * {@link it.pagopa.ecommerce.commons.documents.TransactionRefundedEvent
- * TransactionRefundedEvent}, namely {@link TransactionRefundedData
- * TransactionRefundedData}.
+ * {@link TransactionRefundedEvent TransactionRefundedEvent}, namely
+ * {@link TransactionRefundedData TransactionRefundedData}.
  * </p>
  *
  * @see BaseTransaction
- * @see it.pagopa.ecommerce.commons.documents.TransactionRefundedEvent
+ * @see TransactionRefundedEvent
  */
 @ToString
 @EqualsAndHashCode(callSuper = true)
