+++ src/main/java/it/pagopa/ecommerce/commons/domain/pojos/BaseTransactionWithRequestedAuthorization.java
@@ -1,6 +1,7 @@
 package it.pagopa.ecommerce.commons.domain.pojos;
 
-import it.pagopa.ecommerce.commons.documents.TransactionAuthorizationRequestData;
+import it.pagopa.ecommerce.commons.documents.v1.TransactionAuthorizationRequestData;
+import it.pagopa.ecommerce.commons.documents.v1.TransactionAuthorizationRequestedEvent;
 import lombok.AccessLevel;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
@@ -15,9 +16,9 @@ import lombok.experimental.FieldDefaults;
  * </p>
  * <p>
  * All accrued data is exposed through the same structure generated by
- * {@link it.pagopa.ecommerce.commons.documents.TransactionAuthorizationRequestedEvent
+ * {@link TransactionAuthorizationRequestedEvent
  * TransactionAuthorizationRequestedEvent}, namely
- * {@link it.pagopa.ecommerce.commons.documents.TransactionAuthorizationRequestData
+ * {@link TransactionAuthorizationRequestData
  * TransactionAuthorizationRequestData}.
  * </p>
  */
