+++ src/test/java/it/pagopa/ecommerce/commons/documents/TransactionDocumentTest.java
@@ -1,6 +1,7 @@
 package it.pagopa.ecommerce.commons.documents;
 
 import it.pagopa.ecommerce.commons.TransactionTestUtils;
+import it.pagopa.ecommerce.commons.documents.v1.Transaction;
 import it.pagopa.ecommerce.commons.domain.TransactionActivated;
 import it.pagopa.ecommerce.commons.generated.server.model.TransactionStatusDto;
 import org.junit.jupiter.api.Test;
